change double format to integer
for profileItem type number.
css add elipsis to fullname

+++ grails-app/services/ticbox/RespondentService.groovy
@@ -48,7 +48,7 @@ class RespondentService {
 
                         case ProfileItem.TYPES.NUMBER :
 
-                            profileItems.put(entry.key, Double.valueOf(entry.value))
+                            profileItems.put(entry.key, Integer.valueOf(entry.value))
 
                             break
